refactor(storage): isolate reader lane assembly
Finding: AV1A-B2
